requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Protein and molecular modeling
fair-esm>=2.0.0
rdkit>=2023.3.1
biopython>=1.81

# Chemistry and cheminformatics
chemformer>=1.0.0
selfies>=2.1.1

# Visualization and utilities
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
tqdm>=4.65.0
scipy>=1.11.0

# Reproducibility and configuration
random-seed>=1.0.0
pyyaml>=6.0

# Optional: for enhanced molecular representations
deepchem>=2.7.0